Add EmptySelection key binding mode
With this, a bind can be configured to trigger based on the presence or
absence of a selection.

For example, this config emulates windows terminal style copy paste:
```
[mouse]
hide_when_typing = true
bindings = [
{ mode = "~EmptySelection", mouse = "Right", mods = "Shift", action = "Copy" },
{ mode = "~EmptySelection", mouse = "Right", mods = "Shift", action = "ClearSelection" },
{ mode = "EmptySelection",  mouse = "Right", mods = "Shift", action = "PasteSelection" },
]
```

In the above example, when no text is selected shift-rightclick will
paste. If text is selected, then shift-rightclick will copy and clear
the selection.